requirements.txt
flask
torch
transformers
nltk
numpy
pandas
rouge-score
scikit-learn
sentencepiece
pypdf
python-docx
